Complete limbic system
An attempt at building a complete limbic system. I had to redo the subconscious. Also new, is the separation of time domain for conscious realms. Never considered the possibility until I had to design this schematic. It makes sense, I guess. Once the subconscious detects something, the conscious locks in and only mingles when it reaches the collective subconscious.